refactor: 계층 구조 다시 복구
