+++ app/(deal)/components/BrandProductList.tsx
