+++ app/(deal)/components/DealNavbar.tsx
