+++ app/(deal)/components/DealTimer.tsx
@@ -1,4 +1,4 @@
-import useDealTimer from "@/app/product/(deal)/hooks/useDealTimer";
+import useDealTimer from "@/app/(deal)/hooks/useDealTimer";
 
 function DealTimer() {
   const { timeLeft, progress } = useDealTimer();
