+++ app/(deal)/hooks/useDealTimer.tsx
